+++ src/components/LinkToPatronPreRegistrations/LinkToPatronPreRegistrations.js
@@ -1,0 +1,34 @@
+import { useHistory } from 'react-router-dom';
+import { FormattedMessage } from 'react-intl';
+
+import { IfPermission } from '@folio/stripes/core';
+import {
+  Button,
+  Icon,
+} from '@folio/stripes/components';
+
+
+const LinkToPatronPreRegistrations = () => {
+  const history = useHistory();
+
+  return (
+    <IfPermission perm="ui-users.create">
+      <Button
+        to={{
+          pathname: '/users/pre-registration-records',
+          state: {
+            pathname: history?.location?.pathname,
+            search: history?.location?.search,
+          },
+        }}
+        buttonStyle="dropdownItem"
+      >
+        <Icon icon="search">
+          <FormattedMessage id="ui-users.actionMenu.preRegistrationRecords" />
+        </Icon>
+      </Button>
+    </IfPermission>
+  );
+};
+
+export default LinkToPatronPreRegistrations;