+++ CHANGELOG.md
@@ -43,6 +43,8 @@
 * Mark check box column header as non interactive. Refs UIU-2940.
 * Populate 'State/Prov./Region', 'Zip/Postal Code', 'Country' Fields on user creation via edge module. Refs UIU-3188.
 * Fix issue when slash character in item barcode creates 404 error when adding fee/fine at check in. Refs UIU-3193.
+* Add action button for searching patrons pre registration records. Refs UIU-3219.
+* Display patrons pre registration records results. Refs UIU-3222
 
 ## [10.1.2](https://github.com/folio-org/ui-users/tree/v10.1.2) (2024-09-05)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v10.1.1...v10.1.2)